Fix warning: implicit conversion changes signedness: 'int' to 'unsigned int' [-Wsign-conversion]
visp/example/direct-visual-servoing/photometricMappingVisualServoing.cpp:163:21: warning:
      implicit conversion changes signedness: 'int' to 'unsigned int' [-Wsign-conversion]
  163 |       numDbImages = atoi(optarg_);
      |                   ~ ^~~~~~~~~~~~~
visp/example/direct-visual-servoing/photometricMappingVisualServoing.cpp:166:23: warning:
      implicit conversion changes signedness: 'int' to 'unsigned int' [-Wsign-conversion]
  166 |       numComponents = atoi(optarg_);
      |                     ~ ^~~~~~~~~~~~~

+++ example/direct-visual-servoing/photometricMappingVisualServoing.cpp
@@ -140,7 +140,7 @@ OPTIONS:                                               Default\n\
 
 */
 bool getOptions(int argc, const char **argv, std::string &ipath, bool &click_allowed, bool &display,
-                 int &niter, std::string &method, unsigned &numDbImages, unsigned &numComponents, double &lambda)
+                 int &niter, std::string &method, unsigned int &numDbImages, unsigned int &numComponents, double &lambda)
 {
   const char *optarg_;
   int c;
@@ -160,10 +160,10 @@ bool getOptions(int argc, const char **argv, std::string &ipath, bool &click_all
       method = std::string(optarg_);
       break;
     case 'p':
-      numDbImages = atoi(optarg_);
+      numDbImages = static_cast<unsigned int>(atoi(optarg_));
       break;
     case 'k':
-      numComponents = atoi(optarg_);
+      numComponents = static_cast<unsigned int>(atoi(optarg_));
       break;
     case 'n':
       niter = atoi(optarg_);
